# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/register-for-test": {
      "post": {
        "description": "Register a user account before taking a test. Upload CV and degree files from your local machine. Files will be uploaded to Cloudinary. User will receive credentials after passing the test.",
        "operationId": "Users_registerForTest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "fullName",
                  "email",
                  "phone",
                  "bio",
                  "cv"
                ],
                "properties": {
                  "fullName": {
                    "type": "string",
                    "example": "John Doe",
                    "description": "Full name of the user"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john.doe@example.com",
                    "description": "Email address"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+1234567890",
                    "description": "Phone number in international format"
                  },
                  "bio": {
                    "type": "string",
                    "example": "Experienced software developer with 5+ years in web development",
                    "description": "User bio"
                  },
                  "cv": {
                    "type": "string",
                    "format": "binary",
                    "description": "CV file (PDF format, required)"
                  },
                  "degree": {
                    "type": "string",
                    "format": "binary",
                    "description": "Degree certificate file (PDF format, optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully. You can now take the test.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registration successful. You can now take the test."
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string",
                          "format": "uuid",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "email": {
                          "type": "string",
                          "example": "user@example.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, email already in use, or invalid file format."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Register for test (pre-test registration)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users": {
      "post": {
        "operationId": "Users_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully."
          }
        },
        "security": [
          {
            "JWT-auth": []
          },
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new user (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "Users_findAll",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/register-member-for-club": {
      "post": {
        "description": "Creates/refreshes a pending member applicant account. The applicant must pass the club test before login credentials are emailed.",
        "operationId": "Users_registerMemberForClub",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterMemberForClubDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member applicant registered successfully."
          },
          "409": {
            "description": "Email already belongs to an active account or another account type."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Register a member applicant before club test",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/creator-applications/pending": {
      "get": {
        "operationId": "Users_getPendingCreatorApplications",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending leader applications retrieved successfully."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get pending leader applications for admin review",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "Users_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found."
          },
          "400": {
            "description": "Invalid UUID format."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "Users_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully."
          },
          "400": {
            "description": "Invalid input."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update user by ID",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "Users_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully."
          },
          "400": {
            "description": "Invalid UUID format."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete user by ID",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/memberships": {
      "get": {
        "operationId": "Users_getUserMemberships",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user memberships."
          },
          "400": {
            "description": "Invalid UUID format."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all memberships for a user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/approve-creator": {
      "post": {
        "operationId": "Users_approveCreator",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveCreatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Creator approved."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Approve a creator account",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/reject-creator": {
      "post": {
        "operationId": "Users_rejectCreator",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectCreatorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Creator rejected."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Reject a creator account",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/activate-user": {
      "post": {
        "operationId": "Users_activateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User activation status updated."
          },
          "403": {
            "description": "Forbidden - Admin access required."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Activate or deactivate a user",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "CodeCircle",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Root"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external services"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "user name",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "user last name",
            "description": "Last name of the user"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Optional password"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email"
        ]
      },
      "RegisterMemberForClubDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Jane Doe",
            "description": "Full name of the member applicant"
          },
          "email": {
            "type": "string",
            "example": "jane@example.com",
            "description": "Email of the member applicant"
          },
          "clubId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Target club ID"
          }
        },
        "required": [
          "fullName",
          "email",
          "clubId"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "user first name",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "example": "user last name",
            "description": "Last name of the user"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "Optional password"
          }
        }
      },
      "ApproveCreatorDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab",
            "description": "User ID (UUID) to approve"
          }
        },
        "required": [
          "userId"
        ]
      },
      "RejectCreatorDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab",
            "description": "User ID (UUID) to reject"
          },
          "reason": {
            "type": "string",
            "example": "Insufficient information",
            "description": "Reason for rejection"
          }
        },
        "required": [
          "userId",
          "reason"
        ]
      },
      "ActivateUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "a1b2c3d4-e5f6-7890-abcd-1234567890ab",
            "description": "User ID (UUID) to activate/deactivate"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Set true to activate, false to deactivate"
          }
        },
        "required": [
          "userId",
          "isActive"
        ]
      },
...
    }
  },
...
}